chore(plateform): move browse types to types/api and add hero image
